Rename files, add laws.
