--- control_server/make.bat
@@ -1,1 +1,0 @@
-python make.py py2exe
