+++ control_server/make_for_win.bat
@@ -1,0 +1,1 @@
+python make_for_win.py py2exe
